revert float coords to integers

+++ app/models/user_item.rb
@@ -5,8 +5,8 @@ class UserItem
   embedded_in :user
   belongs_to :item
 
-  field :x,                 type: Float,  default: -1
-  field :y,                 type: Float,  default: -1
+  field :x,                 type: Integer,  default: -1
+  field :y,                 type: Integer,  default: -1
   field :status,            type: String, default: 'standby'
   field :room_id,           type: String
   field :rotation,          type: Integer
